+++ src/pages/offices/components/offices/offices.component.js
@@ -1,0 +1,19 @@
+const speaker = require("../../../../util/speaker");
+
+module.exports = {
+    onCreate(input) {
+        this.state = {
+            voices: []
+        };
+    },
+
+    onMount() {
+        $(function () {
+            speaker.loadVoices((voices) => {
+                if (voices && voices.length > 0) {
+                    this.state.voices = voices;
+                }
+            });
+        }.bind(this));
+    }
+}