+++ src/pages/players/components/players/players.component.js
@@ -1,11 +1,14 @@
 const io = require(`../../../../util/socket.io-helper`);
+const localStorage = require("../../../../util/localstorage");
 const _ = require("underscore");
+const speaker = require('../../../../util/speaker');
 
 module.exports = {
     onCreate(input) {
         this.state = {
             players: input.players,
-            smartcardReadhFnc: (data) => { }
+            smartcardReadhFnc: (data) => { },
+            ttsVoice: undefined
         }
     },
     onMount() {
@@ -17,6 +20,16 @@ module.exports = {
             const comp = this.getComponent('add-player-modal');
             this.state.smartcardReadhFnc = comp.smartcardRead.bind(comp);
         }.bind(this));
+
+        const venueId = localStorage.get('venue_id');
+        if (venueId) {
+            let venues = this.input.venues.filter((venue) => venue.id == venueId);
+            if (venues.length > 0) {
+                this.state.ttsVoice = venues[0].config.tts_voice;
+            }
+        }
+        // Initialize voices
+        speaker.loadVoices(() => {});
     },
     officeChanged(id) {
         const officeId = parseInt(id);